api/requirements.txt
fastapi
python-dotenv
mysql-connector-python
uvicorn